Introduce Layout and Lint cops as shared config
This is the start of work to define shared rubocop config for
FutureLearn ruby projects. This provides two advantages:

* We can ensure consistent rules in our different repos as much as
  possible

* It provides a central place to document the rationales for when we
  diverge from rubocop defaults

In future I'd like to see if we can generate an HTML style guide from
this config that can be referenced from our main documentation.

This commit kicks things off by adding our config for the Layout and
Lint namespaces. These contain by and large the least controversial
cops. It also disables the Metrics namespace, which we don't make any
use of.

I've imported these directly from the futurelearn repo. I've also added
comment rationales for all of our config, so that we can use these files
as documentation for where we've diverged from the defaults. Where we're
open to enabling (or actively seeking to enable) a cop in future, I've
noted that. This doesn't imply that others are set in stone, but I
thought where the decisions are a real coin toss, it's good to reflect
that.