Add rubocop-thread_safety config
Since our main app runs in puma which is threaded, we use
rubocop-thread_safety for some basic linting advice.

We don't have any exclusions for this plugin's rules, because they're
all broadly valuable. In rubocop terms they're mostly "unsafe" because
to pass they typically require changes that affect application
behaviour, but in this case that's actually desirable. We should pair
this with a guidelines section on thread-safe coding.