Add rubocop-graphql config
We recently started using rubocop-graphql as a first attempt at bringing
some consistency to how write GraphQL code.

Now that we have a separate repo for managing our rubocop config this
feels like the natural place for this config.

There are a number of cops which have been overridden, most of which
are cops which enforce descriptions for everything. I believe that
descriptions are valuable when something needs additional explanation,
however I don't think we should be enforcing them for everything given
that this will lead to a lot of unnecessary descriptions for example:

```ruby
class UserType < BaseType
  field :id, GraphQL::Types::ID, description: "ID of the user", null: false
end
```