Set Style/HashSyntax EnforcedShorthandSyntax to 'either'
When we upgrade to ruby 3.1 we'll have a great many candidates for
hashes that could use the new shorthand { foo:, bar: } hash literal
syntax.

To manage the migration we'll want to first get on ruby 3.1 and only
then think about making non-essential syntax changes, so this sets the
cop to be permissive for the time being.